@@ -1,18 +1,20 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import App from "./component/app/app";
 import theme from "./service/theme";
-import { ColorStateProvider } from "./context/color-state-context";
-import { CssBaseline, ThemeProvider } from "@mui/material";
+import {ColorStateProvider} from "./context/color-state-context";
+import {CssBaseline, ThemeProvider} from "@mui/material";
+import {createRoot} from "react-dom/client";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      <ColorStateProvider>
-        <App />
-      </ColorStateProvider>
-    </ThemeProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
+const container = document.getElementById("root");
+const root = createRoot(container!);
+
+root.render(
+    <React.StrictMode>
+        <ThemeProvider theme={theme}>
+            <CssBaseline />
+            <ColorStateProvider>
+                <App />
+            </ColorStateProvider>
+        </ThemeProvider>
+    </React.StrictMode>
 );
